+++ packages/analytics-js-common/CHANGELOG.md
@@ -2,6 +2,13 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [3.24.0](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js-common@3.23.0...@rudderstack/analytics-js-common@3.24.0) (2025-09-01)
+
+
+### Features
+
+* add dmt support for cloned destinations ([#2434](https://github.com/rudderlabs/rudder-sdk-js/issues/2434)) ([e701d6f](https://github.com/rudderlabs/rudder-sdk-js/commit/e701d6fb81edf1c326f1e50c4f1a040016521cd8))
+
 ## [3.23.0](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js-common@3.22.0...@rudderstack/analytics-js-common@3.23.0) (2025-08-11)