+++ packages/analytics-js/CHANGELOG.md
@@ -2,6 +2,13 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [3.23.2](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js@3.23.1...@rudderstack/analytics-js@3.23.2) (2025-09-01)
+
+### Dependency Updates
+
+* `@rudderstack/analytics-js-cookies` updated to version `0.5.2`
+* `@rudderstack/analytics-js-common` updated to version `3.24.0`
+* `@rudderstack/analytics-js-plugins` updated to version `3.13.0`
 ## [3.23.1](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js@3.23.0...@rudderstack/analytics-js@3.23.1) (2025-08-27)
 
 ### Dependency Updates