+++ packages/analytics-v1.1/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [2.52.4](https://github.com/rudderlabs/rudder-sdk-js/compare/rudder-sdk-js@2.52.3...rudder-sdk-js@2.52.4) (2025-09-01)
+
+### Dependency Updates
+
+* `@rudderstack/analytics-js-integrations` updated to version `3.18.0`
 ## [2.52.3](https://github.com/rudderlabs/rudder-sdk-js/compare/rudder-sdk-js@2.52.2...rudder-sdk-js@2.52.3) (2025-08-27)
 
 ### Dependency Updates